requirements.txt
pyside6==6.7.3

winrt-windows.data.xml.dom==2.2.0
winrt-windows.ui.notifications==2.2.0
